Completed sitemap implementation
